Lab 7.3 : Generalize DecimalCalculator and BinaryCalculator by creating BaseCalculator
and edit some test case. SORRY
